Basic Projectile Infrastructure Started
Changed name attackPrefab to spellPrefab in order to allow more generalization of spell effects in the future. Created projectileManager script for general control over projectile spells, added fireball prefab and animation, and added an instantiate method in HandManager to spawn a spell on play which is then handled independently of HandManager.