Added Object Spawning and Custom Scripts
Once this is pushed to main, anyone can edit the settings of the biomes, including the objects they spawn, and even add custom code to be run during generation. Click the "Map" in the top left, and change info in the inspector. Right now, the system randomly spawns any object entered in the decorObjects list on tiles of that biome at the entered percentage chance on each tile.